tweak: move subscribe to its own component
